+++ plans/apps/export-posters/src/day-types.ts
@@ -9,7 +9,6 @@ export const DAY_TYPES: DayTypeConfig[] = [
 
 	{
 		_id: 'DT_ESC_DU',
-		dates: [],
 		day_type: '1',
 		index: 1,
 		name: 'Escolar | Dias Úteis',
@@ -17,7 +16,6 @@ export const DAY_TYPES: DayTypeConfig[] = [
 	},
 	{
 		_id: 'DT_ESC_SAB',
-		dates: [],
 		day_type: '2',
 		index: 2,
 		name: 'Escolar | Sábados',
@@ -25,7 +23,6 @@ export const DAY_TYPES: DayTypeConfig[] = [
 	},
 	{
 		_id: 'DT_ESC_DOM',
-		dates: [],
 		day_type: '3',
 		index: 3,
 		name: 'Escolar | Domingos e Feriados',
@@ -36,7 +33,6 @@ export const DAY_TYPES: DayTypeConfig[] = [
 
 	{
 		_id: 'DT_FER_DU',
-		dates: [],
 		day_type: '1',
 		index: 4,
 		name: 'Férias Escolares | Dias Úteis',
@@ -44,7 +40,6 @@ export const DAY_TYPES: DayTypeConfig[] = [
 	},
 	{
 		_id: 'DT_FER_SAB',
-		dates: [],
 		day_type: '2',
 		index: 5,
 		name: 'Férias Escolares | Sábados',
@@ -52,7 +47,6 @@ export const DAY_TYPES: DayTypeConfig[] = [
 	},
 	{
 		_id: 'DT_FER_DOM',
-		dates: [],
 		day_type: '3',
 		index: 6,
 		name: 'Férias Escolares | Domingos e Feriados',
@@ -63,7 +57,6 @@ export const DAY_TYPES: DayTypeConfig[] = [
 
 	{
 		_id: 'DT_VER_DU',
-		dates: [],
 		day_type: '1',
 		index: 7,
 		name: 'Verão | Dias Úteis',
@@ -71,7 +64,6 @@ export const DAY_TYPES: DayTypeConfig[] = [
 	},
 	{
 		_id: 'DT_VER_SAB',
-		dates: [],
 		day_type: '2',
 		index: 8,
 		name: 'Verão | Sábados',
@@ -79,7 +71,6 @@ export const DAY_TYPES: DayTypeConfig[] = [
 	},
 	{
 		_id: 'DT_VER_DOM',
-		dates: [],
 		day_type: '3',
 		index: 9,
 		name: 'Verão | Domingos e Feriados',
