+++ plans/apps/export-posters/src/types.ts
@@ -37,7 +37,6 @@ export interface DayTypesExt {
 
 export interface DayTypeConfig {
 	_id: string
-	dates: OperationalDate[]
 	day_type: '1' | '2' | '3'
 	index: number
 	name: string
